add threading support to performance tests
